# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  total_items: number;
  total_stock: number;
  low_stock_count: number;
  out_of_stock_count: number;
  stock_in_this_month: number;
  stock_out_this_month: number;
}