# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemUpdateStatus } from './itemUpdateStatus';

export interface ItemUpdate {
  item_code?: string;
  item_name?: string;
  spec?: string;
  category?: string;
  unit?: string;
  location?: string;
  min_stock?: number;
  minimum_order?: number;
  opening_stock?: number;
  photo_url?: string;
  status?: ItemUpdateStatus;
}