# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileRole } from './profileRole';
import type { ProfileStatus } from './profileStatus';

export interface Profile {
  id: number;
  clerk_user_id: string;
  email: string;
  /** @nullable */
  full_name?: string | null;
  role: ProfileRole;
  status: ProfileStatus;
  created_at: Date;
}