# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementMovementType } from './stockMovementMovementType';

export interface StockMovement {
  id: number;
  item_id: number;
  /** @nullable */
  item_code?: string | null;
  /** @nullable */
  item_name?: string | null;
  movement_type: StockMovementMovementType;
  quantity: number;
  movement_date: Date;
  /** @nullable */
  reference_no?: string | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  requested_by?: string | null;
  /** @nullable */
  issued_by?: string | null;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  remarks?: string | null;
  created_by: string;
  created_at: Date;
}